I hereby agree to the terms of the [GreptimeDB CLA](https://gist.github.com/xtang/6378857777706e568c1949c7578592cc)

## What's changed and what's your intention?


### Result

Up to 50% latency cut in tsbs scenario

| test                     | main(ms)   | filter(ms) | opt. rate |
|--------------------------|------------|------------|-----------|
|   **cpu-max-all-1***          |   18.20    |   14.13    |   **0.78**    |
|   cpu-max-all-8          |   81.57    |   79.57    |   0.98    |
|   **double-groupby-1***       |   749.72   |   728.76   |   0.97    |
|   double-groupby-5       |   1156.25  |   1193.33  |   1.03    |
|   double-groupby-all     |   1925.41  |   1902.98  |   0.99    |
|   groupby-orderby-limit  |   453.77   |   499.06   |   1.10    |
|   **high-cpu-1***             |   19.01    |   9.69     |   **0.51**    |
|   high-cpu-all           |   6189.08  |   6402.76  |   1.03    |
|   lastpoint              |   6179.84  |   6141.70  |   0.99    |
|   **single-groupby-1-1-1***   |   11.91    |   10.51    |   **0.88**    |
|   **single-groupby-1-1-12***  |   13.97    |   8.58     |   **0.61**    |
|   single-groupby-1-8-1   |   22.24    |   20.39    |   0.92    |
|   **single-groupby-5-1-1***   |   10.29    |   9.26     |   **0.90**    |
|   **single-groupby-5-1-12***  |   15.30    |   10.00    |   **0.65**    |
|   single-groupby-5-8-1   |   30.01    |   30.37    |   1.01    |

**are featured cases*

### Explain

This patch adds a `SimpleFilterEvaluator` struct that can calculate filter expr in the simple form. See its doc for details.

By using `SimpleFilterEvaluator` in parquet reader, it now can filter out invalid rows in advance. The phase it takes effort is right after a batch is retrieved from parquet file. Thus can reduce the consumption in later process phases like sort / dedup etc.

This also fixes a possible bug when an empty batch is read from parquet file, see UT case `test_read_empty_batch`.

## Checklist

- [x]  I have written the necessary rustdoc comments.
- [x]  I have added the necessary unit tests and integration tests.
- [x]  This PR does not require documentation updates.

## Refer to a related PR or issue link (optional)